Removed paper node module from gitignore
I have made a change to the acorn library used in the paper node module
that fixes an error. The paperjs maintainers are aware of the issue and
will investigate if updating to the latest version of acorn fixes it or
they will go ahead and implement my fix into the paperjs repo.